// src/interfaces/ComplexesSchedules.tsx
// Generated by https://quicktype.io

export interface PutMultipleSchedulesResponse {
  updatedComplexSchedules: ComplexSchedule[];
}
export interface PostMultipleSchedulesResponse {
  newComplexSchedules: ComplexSchedule[];
}
export interface DeleteScheduleResponse {
  removedSchedules: string[];
}

export interface ComplexSchedule {
  _id: string;
  weekDays: number[];
  openingTime: string;
  closingTime: string;
  sport: string;
  complexId: string;
  updatedAt: string;
  createdAt: string;
}